+++ django/core/adaptors/base_adaptor.py
