--- django/cron_tasks/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CronTasksConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "cron_tasks"
